+++ src/config/firebase/collections.js
@@ -1,0 +1,2 @@
+export const USERS = "users";
+export const STAMPS = "stamps";
