+++ src/index.js
@@ -3,20 +3,17 @@ import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App";
-// import { AuthProvider } from "./contexts/AuthContext";
 import reportWebVitals from "./reportWebVitals";
-import "./utils/firebase";
+import "./config/firebase";
 import store from "./app/store.js";
 
 ReactDOM.render(
   <React.StrictMode>
-    {/* <AuthProvider> */}
     <Provider store={store}>
       <BrowserRouter>
         <App />
       </BrowserRouter>
     </Provider>
-    {/* </AuthProvider> */}
   </React.StrictMode>,
   document.getElementById("root")
 );
